[BUGFIX] Makefiles: parallel make (-jn) fails as *.pb.h files missing
Parallel make runs all targets at once, the targets the require
onnx.pb.* files fail if the target that generate those files (target
named 'onnx') has'nt been called yet. make being oblivous to this fact
exits with an error. A solution to this is to explicityly declare onnx
as the dependencies of all the targets. This is messy. Current solution
takes onnx.pb.* generation out of the make file into build.sh and calls
make from within.

This is a terrible solution (hack), will have to be looked into later
(perhaps use cmake or autotools).

Signed-off-by: Shreeyash Pandey <shreeyash335@gmail.com>