onnx_parser: change connection logic for nodes
earlier, output_map stored {node.name, node->next.vertex}. this
resulted in in-correct parsing of nodes with more than one
inputs/outputs.

current logic overcomes this by storing the {node.outputs.name(),
node.vertex} in the output_map in the add stage. In the connection
stage, for current_node, nodes corresponding to its inputs are
queried from the output_map and connected to the current node.

Signed-off-by: Shreeyash Pandey <shreeyash335@gmail.com>