fix(sasa,tensor): adder omits single channeled outputs
adder was omiting single channeled outputs as the second-inner for loop
checked from i = 0 to size - 1 which resulted in i = 0 to 0. no outputs
were generated. this patch fixes the bug by moving to a cleaner design.

TensorExtant::at() deduces the type of underlying tensor and where
data could be present by the type assumed by the user. Better to
deduce it from the TensorProto.data_type() as raw_data may be used
to store any datatype.

also adds comments to functions.

Signed-off-by: Shreeyash Pandey <shreeyash335@gmail.com>