fix(instgen): fix FC-output kernel_itr mismatch
the value that kernel_itr takes in the output inst following an FC
instruction is decided on the basis of true COLS of the weight matrix
and the size of the vector array.

weight matrix may be present in the form of AxB or BxA (depending on the
transB param), the true dim is found by applying this transposition.
cols of the true dimensions are supposed to be used for calculating
kernel_itr. this was not being done, hence the bug.

Signed-off-by: Shreeyash Pandey <shreeyash335@gmail.com>