fix(rt): remove hasheddispatchtable
HDT was a solution to a problem that occurred when compile and run
were separate processes. Since the new python interface, both run
in the same process. As a result, dispatch args specified for compile
are available for run too. This makes reading dispatch args from the
GML not necessary. Therefore, we can do away with HDTs

Signed-off-by: Shreeyash Pandey <shreeyash335@gmail.com>